requirements.txt
dnspython==2.2.1
pykerberos==1.2.4
pymongo==4.1.1
python-dotenv==0.20.0
python-snappy==0.6.1
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.7.3
pydantic==1.9.1
pykerberos==1.2.4
pymongo==4.1.1
python-dotenv==0.20.0
python-multipart==0.0.5
python-snappy==0.6.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
